pyproject.toml
[project]
name = "ttb-cola-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.0",
    "streamlit>=1.45.1",
    "python-dotenv>=1.0.0",
]
